src/api/requirements.txt
aiofiles==24.1.0
celery==5.5.1
fastapi==0.115.12
flower==2.0.1
pytest==7.2.2
redis==5.2.1
requests==2.28.2
uvicorn==0.34.2
python-multipart==0.0.9
rio_cogeo==5.4.1
fastapi-keycloak-middleware==1.2.0
python-dotenv==1.1.0
cryptography==3.4.7